Fix 4 conversion arguments to var_desc calls
  Corrected 4 conversion arguments in calls to var_desc for temperatures and
salinities, so that they are consistent with the units of these variables and
the described purpose of the conversion element of the var_desc type.  Until the
conversion arguments to modify_vardesc and query_vardesc, these incorrect values
were inconsequential, but now they need to be fixed before they are
inadvertently used.  All answers are bitwise identical.