Rescale strat_floor
  Rescaled the internal representation of the FGNV_STRAT_FLOOR variable and
thickness_diffuse%N2_floor to include appropriate factors of US%Z_to_L to
reflect the scaling of aspect ratios.  This leads to the elimination of one
rescaling factor in thickness_diffuse_full() and its replacement by another
in a scale factor for a get_param call.  All answers are bitwise identical.