Diagnose area integrated fluxes in rescaled units
  Keep 36 area integrated surface mass, heat or salt flux diagnostics in
forcing_diagnostics in rescaled units by replacing an unscale argument with a
tmp_scale argument to global_area_integral.  Also added the corresponding
conversion arguments to the register_scalar_field calls for each of these
diagnostics, so that the documented units and conversion factors can be used to
confirm the correctness of the documented units for these variables.  The
internal variables used for the integrated or averaged heat, mass or salt fluxes
were also separated for clarity about the units of these variables.  All answers
and diagnostics are bitwise identical.