+Refactor shelfwave OBC code
  Refactored the shelfwave OBC code to correct a number of minor problems.  It
corrects some dimensionally inconsistent expressions. It adds a rescaling factor
to ensure that the solutions are mathematically equivalent for various units
used to describe the Cartesian horizontal grid, and adds a fatal error if the
grid is not Cartesian.  It eliminates several redundant hard-coded dimensional
values used to initialize the control structure.  It eliminates elements from
the control structure for this type that are not reused after initialization.
Finally, it adds the new runtime parameter SHELFWAVE_CORRECT_AMPLITUDE that
causes the inflow velocity amplitude to actually match the value specified in
SHELFWAVE_AMPLITUDE.  When SHELFWAVE_CORRECT_AMPLITUDE is true, it alters the
default value of SHELFWAVE_AMPLITUDE to give roughly similar amplitudes of the
flow to those with the default value when it is false.  By default solutions are
unchanged, but there is a new runtime parameter.