Add compressibility to linear equation of state
This commit is part 1/2 of enabling user specified compressibility in
linear equation of state. An attribute drho_dp is added in extended type
linear_EOS and various calculations for density, specific volume and
their first order derivatives. But the full compressibility is not yet
added in density and specific volume integrals.

* Two recently added runtime parameters (TREF and SREF) for linear EOS
are renamed to T_REF_LINEAR_EOS and S_REF_LINEAR_EOS, respectively. The
old names could be confused with T_REF and S_REF, which are used to
initialize the model's temperature and salinity in certain methods.
RHO_TREF_SREF is also renamed to RHO_REF_LINEAR_EOS.
* Description of parameter RHO_T0_S0 is modified for clarification.
* Two new runtime parameters are added (DRHO_DP and P_REF_LINEAR_EOS).
* `Compressible` attribute of linear_EOS type is deterimined by whether
DRHO_DP is zero or not. (At the moment, the attribute is only used to
prevent user from using Montgomery PGF when EOS is compressible.)