Simpler btloop_eta_predictor
  Simplified btloop_eta_predictor by moving several optional loops out into the
main body of btstep_timeloop, to increase the chances that this routine will
become a good candidate for inlining, and to take steps toward the reuse of the
code calculating the barotropic fluxes in the predictor and corrector steps. In
addition, several unused variables were eliminated, and the arguments to
btloop_eta_predictor are now documented in the order in which they appear in the
argument list.  All answers are bitwise identical and no public interfaces are
changed.