Refactor barotropic OBC code
  Refacted the code implementing the barotropic open boundary conditions to
replace the direct references to the segments with simple 2-d arrays in the
BT_OBC type for everything inside of the performance-critical routine
btstep_timeloop.  There are several new or renamed elements inside of the
BT_OBC_type and three new integer parameters enumerating the types of open
boundary conditions that are to be applied at each point, with the sign of
integers in the new u_OBC_type and v_OBC_type arrays indicating the direction of
the OBCs, and 0 indicating no OBC at a cell face.  There is a new routine,
initialize_BT_OBC, that stores the static information about the position and
nature of the various OBCs for use within the barotropic time stepping; this
routine is called from barotropic_init.  Several spelling errors in comments
were also corrected.  All answers are bitwise identical and no public interfaces
are changed.