Refactor btcalc to reduce duplicated code
  Refactored btcalc to include the application of open boundary conditions in
the calculation of CS%frhatu and CS%frhatv within the same loops as the
non-OBC calculations, and also to reduce the amount of duplicated code.  This
should increase the model efficiency with open boundary conditions, and have
minimal performance impacts otherwise.  All answers are bitwise identical
and no public interfaces are changed.