+(*)Arbitrary grid rotation with MOM_open_boundary
  Revised MOM_open_boundary to better handle grid rotation for an arbitrary
number of turns.  This includes adding the new routine write_OBC_info() that can
be used to write extensive information about the settings and the contents of
arrays in the ocean_OBC_type.  Because write_OBC_info() is rather verbose, it
should probably only be used for debugging.  Also added the new routine
allocate_rotated_seg_data() and use it in place of allocate_rotated_array() in
rotate_OBC_segment_data() to avoid allocating arrays with the wrong index ranges
(but of the correct size).  Also modified rotate_OBC_segment_data to copy over
information in the tr_Reg elements of the segment when they are in use.  Prior
to this change, there would be a segmentation fault if the segment tracer
registry was in use and the index rotation was being enabled (even if the grid
was not actually being rotated).  All answers in non-rotated cases are bitwise
identical, and some rotated cases that previously gave segmentation faults are
new running, although rotation does not yet work for all cases with open
boundary conditions.