+(*)Fix bugs with dyed_channel_update_flow
  When there are specified open boundary conditions, dyed_channel_update_flow()
was setting the normal transport with the wrong units of [L2 T-1 ~> m2 s-1]
instead of [H L2 T-1 ~> m3 s-1 or kg s-1], effectively omitting the layer
thickness in the transports.  This has been corrected by adding the new runtime
parameter CHANNEL_FLOW_OBC_TRANSPORT_BUG that either uses a fixed thickness if
true (of 1 m or 1 kg m-2) or the actual interior thickness.  In either case, the
model is now passing the dimensional consistency testing.  This change requires
a new layer thickness argument to  dyed_channel_update_flow().  In addition,
code was added to dyed_channel_update_flow so that it also satisfies rotational
consistency.  By default, answers are bitwise identical when no dimensional
rescaling or grid rotation is applied, but there is a new runtime parameter and
a new argument to dyed_channel_update_flow.