+Add chksum_OBC_segments
  Added the new publicly visible subroutine chksum_OBC_segments() for debugging
open boundary conditions, and call this new routine via write_OBC_info().
Chksum_OBC_segments() does a checksum on all allocated elements of the open
boundary condition segments and optionally writes out a number of layers of
segment data, with the order of output chosen to appear as it would if the data
were for an eastern open boundary.  There are also modifications to
write_OBC_info() to compensate for grid rotation so that the output should be
largely invariant to rotation.  This commit also adds scalar_pair arguments to 4
uvchksum calls for radiation and oblique open boundary conditions.  All
solutions are bitwise identical, but there is a new publicly visible interface
for debugging checksums.