+(*)Add OBC_HOR_INDEXING_BUG and correct rotation
  Added the new runtime parameter OBC_HOR_INDEXING_BUG that can be set to true
to retain previous answers or false to correct 5 horizontal indexing bugs that
prevent solutions with open boundary conditions from satisfying rotational
symmetry.  Another 3 bugs that only apply to the ability to rotate the model but
do not change answers in the unrotated model were fixed directly without there
being a runtime flag to recreate the old (wrong) answers.  Most of these bugs
are related to the use of the brushcutter mode to set various terms in the open
boundary conditions, but others the use of remapping boundary condition fields
used to specify the tangential flow.  Other bugs that were fixed in this commit
relate to the ability to use a restart with OBCs and a rotated state or
rotationally invariant checksums, or the ability to read input files onto a
rotated grid.  By default, all answers are bitwise identical, but there is a new
runtime parameter in some MOM_parameter_doc files.