+Add rotate_OBC_segment_direction
  Added the new function rotate_OBC_segment_direction that returns an integer
indicating the direction of an OBC segment on rotated grid, given its direction
on the current grid and the number of turns by which that grid has been
rotated.  A negative number of turns undoes the rotation.  For now this new
routine is used in rotate_OBC_segment_config, but additional uses will follow
with subsequent commits.  All answers are bitwise identical, but there is a new
publicly visible function.