Refactoring for OBC restart fix
  This commit makes a set of changes to refactor code related to open boundary
conditions to correct issues with restarts.

  Renamed open_boundary_init to open_boundary_halo_update to reflect what it
actually does and eliminated the 4 unused arguments to this routine.

  Moved halo updates for temperature and salinity and the call to
setup_OBC_tracer_reservoirs out of fill_temp_salt_segments and made the tv
argument to fill_temp_salt_segments intent in.

  Moved the call to setup_OBC_tracer_reservoirs and the debugging calls for OBCs
into the same code block as the rest of the OBC code in MOM_initialize_state.

  Moved the call to open_boundary_halo_update after MOM_state_initialization for
both rotated and unrotated branches.

  Moved the calls to fill_temp_salt_segments and setup_OBC_tracer_reservoirs and
the debugging calls for OBCs into the same code block as the rest of the OBC
code in MOM_initialize_state.

  All answers are bitwise identical, but there are changes to public interfaces
and what does and does not occur in several routines.