+Add calc_isoneutral_slopes optional argument OBC_N2
  Added the new optional argument OBC_N2 to calc_isoneutral_slopes that can be
use to specify that only interior points are used to calculate the buoyancy
frequencies at velocity points when open boundary conditions are in use.  When
the new OBC_N2 argument is absent or false, the solutions are identical to what
they had been previously, but this gives a dependence on the arbitrary outer
thicknesses when open boundary conditions are used.  By default, all answers are
bitwise identical, but there is a new optional argument to a publicly visible
interface.