+Add OBC_RESERVOIR_RESTART_BUG to fix OBC restarts
  This commit revises setup_OBC_tracer_reservoirs and other OBC routines to
allow some cases using open boundary conditions and tracer reservoirs to
reproduce across restarts and added the new runtime parameter
OBC_RESERVOIR_RESTART_BUG that can be set to false to enable these corrections
to the OBC restart capabilities.   There is a new optional MOM_restart_CS
argument to setup_OBC_tracer_reservoirs that can be used to query whether the
reservoirs have been read from a restart file before resetting them to values
that may have been set during initialization.  If the reservoirs have not been
set from a restart file, there is now a preference to set them from the tracer
reservoir (tres) element in the OBC tracer registry on the segments (if it has
been initialized) rather than the interior tracer values that have been recorded
on the segment.

  Fill_temp_salt_segments and fill_obgc_segments have been modified to only set
the tres segment data if it has not been initialized already, as indicated by
the is_initialized field.

  This commit adds the new public routine set_initialized_OBC_tracer_reservoirs
that can be used to record in the restart control structure that the OBC%tres_x
and OBC%tres_y fields have been initialized and should not be reset to new
values during initialization.

  The calls to setup_OBC_tracer_reservoirs and open_boundary_halo update have
been moved after the call to tracer_flow_control_init so that the full arrays of
tracer reservoirs do not need to be initialized in fill_obgc_segments.
Fill_obgc_segments was refactored to separate out the steps that should occur
there from those that should not and are wrapped in a bug flag.

  A debugging checksum call to chksum_OBC_segments was added to the top of
radiation_open_bdry_conds to assist in any future debugging of the OBC code.

  Added doxygen comments describing several routines that were previously
missing such a description.

  By default, all answers are bitwise identical, but there is a new publicly
visible routine, a new optional argument to a public interface and a new runtime
parameter that can be used to correct the previous behavior that broke the
reproducibility across restarts of some cases using open boundary conditions
with tracer reservoirs.