vertvisc: Rewrite vertvisc_remnant() to kji loops
As with vertvisc(), this patch rewrites the vertvisc_remnant()
tridiagonal solvers to run in kji order, with even greater benefits to
runtime.  Three instances are shown below.  Speedup is about 1.3-1.4x.

* mom_vert_friction_mp_vertvisc_remnant_:   0.939s,   1.241s (-24.3%)
* mom_vert_friction_mp_vertvisc_remnant_:   0.935s,   1.265s (-26.0%)
* mom_vert_friction_mp_vertvisc_remnant_:   0.910s,   1.258s (-27.7%)

As before, only the diagnoal array (b1) was promoted to 3d.

As with vertvisc() this change is expected to be highly favorable to
GPU performance.