GitHub: FMS-MOM6 build jobs, tarball artifacts
Inconsistencies in jobs within a workflow were leading to differences in
compiler versions across jobs.  This was breaking GitHub Actions
workflows.

We want to continue using `*-latest` versioning, so this patch merges
the FMS build into each MOM6 build.  Although there is greater compute
time, it should be happening in parallel.  It may even be shorter, since
we do not need to wait on the additional OS setup.

This patch also packs executables and other data into tarballs.  This
preserves most of the file data (file mode permissions and
modified-time, mainly) so that we don't have to do Makefile tricks or
use chmod to get things working in the next job.  We also don't need to
resort to horrible awk tricks to generate executable lists.

There may be a slight speedup here if the builds happen in parallel,
since we are doing one less image setup, but CI runtimes are very
variable so I don't think we'll be able to know for now.