+Add MASK_COASTAL_PRESSURE_FORCE runtime parameter
  Add the new runtime option MASK_COASTAL_PRESSURE_FORCE to use land masks to
zero out the diagnosed barotropic pressure gradient accelerations at coastal or
land points.  If it is enabled, this changes diagnostics and it improves the
reproducibility of certain debugging checksums, but it does not alter the
solutions themselves.  By default, all answers and diagnostics are bitwise
identical, but there is a new runtime parameter in some MOM_parameter_doc files.