Use OBC loop ranges on PE in 4 modules
  Refactored the OBC-related code in 5 routines (set_viscous_BBL(),
vertvisc_coef(), find_coupling_coef(), calc_isoneutral_slopes() and
ALE_remap_set_h_vel_OBC()) in 4 modules to use the restricted OBC loop ranges
for the various orientations of OBCs for greater efficiency, and in some cases
to avoid the use of multiple levels of nested types and looping over segments
that seem to be causing problems with our efforts to port MOM6 to GPUs.  The
cases where the loop over the OBC segments continue to be used are those where
the details of the segments matter (and not just their orientation) or where
they are being applied in ways where the order of the segments might matter.
All answers are bitwise identical, and there are no changes to any publicly
visible interfaces.