+Do not call write_ocean_geom from MOM_init_fixed
  Removed the option to call write_ocean_geometry_file() from within
MOM_initialize_fixed().  This option was only being exercised in one of the
three places where MOM_initialize_fixed was being called, and by doing so it
precluded the possibility of specifying an alternative name for this file.
Removing this option eliminates two arguments to MOM_initialize_fixed(), but it
requires the addition of a new call to write_ocean_geometry_file() directly from
shelf_main in ice_shelf_driver.F90.  All answers are bitwise identical, but
there is a simplification in the arguments to a publicly visible routine.