+Rename and split rotate_OBC_init()
  Split rotate_OBC_init() into rotate_OBC_segment_fields() and
rotate_OBC_segment_tracer_registry(), and also split out
rotate_OBC_segment_tracer_data() from rotate_OBC_segment_data(), so that the
segment tracer registry can be handled separately and at different points in the
code from the other information on the OBC segment types.  In addition, the
calls to these two new interfaces have been moved much earlier in initialize_MOM
to sit next to other related OBC code.  This is being done in anticipation that
the tracer registries need not be dealt with on the unrotated OBC type when
there is grid rotation and that this will lead to a simplification of the OBC
code and the calls to it.  All answers are bitwise identical, but there are
changes to the public interfaces to the OBC code.