Deallocate OBC memory with open_boundary_end
  Added calls to open_boundary_end for the potentially rotated OBC type in
MOM_end() and for the unrotated OBC type after the last time it is used in
initialize_MOM().  The former is standard clean-up at the end of a run, while
the latter will avoid a one-time memory leak when the model is run with
grid-rotation.  All answers are bitwise identical.