+Add OBC_REVERSE_SEGMENT_ORDER
  Added the new debugging runtime parameter OBC_REVERSE_SEGMENT_ORDER that can
be set to true to internally store and apply the OBC segments in the reverse of
the order with which they are specified by the used.  When the OBC segments are
order independent, changing this value should give bitwise identical results.
There is a new integer argument indicating the segment to use for setting
parameters in setup_u_point_obc and setup_v_point_obc, but these are internal
routines that are not publicly visible.  This commit also changes some of the
segment index variables from l (which can be hard to distinguish from 1) to n,
and it standardizes the spacing in some loops over segments or in variable
assignments to better align with the MOM6 style guide.  By default the new
runtime parameter is false and all answers are bitwise identical.