+Optionally avoid projecting data for OBCs
  Adds the option to avoid using exterior data at open boundary condition
points, and to avoid projecting information outside across open boundary
condition faces.  Such projections are problematic because there can be pairs
of OBCs that would project inconsistent values into the same point.  This
option is enabled by setting the new runtime parameter OBC_PROJECTION_BUG =
False. To implement these changes, the weights for the barotropic solver to use
when interpolating total thickness onto vorticity points when calculating the
barotropic potential vorticity for use in the Coriolis scheme are now being
stored in the new q_wt element in the barotropic control structure.  These
weights are set up during initialization with information about open boundary
conditions optionally taken into account.  In addition the total thicknesses at
open boundary velocity points that are combined with the potential vorticities
to give the barotropic Coriolis terms are always taken from the interior of an
OBC face.  Some inconsistencies in the units of a negligible but positive cell
volume in the denominator of the barotropic PVs were also corrected in a way
that will not change answers when dimensional rescaling is not being applied.

  This commit also includes the run-time configurable option to set the corner
point land-mask at the bay-like intersection of open boundary conditions to be
unmasked so that open boundary conditions can be applied there.

  By default all answers are bitwise identical, but answers in cases with open
boundary conditions do change when these new capabilities are enabled.