makedep: Update interpreter directive to python3
The interpreter directive ("shebang") of makedep is updated to
`python3`, rather than the version-agnostic `python`.

Although we never invoke the shebang of the script, there are OS
environments out there which will object to any presence of a
versionless python.  PEP 394 also strongly recommends the adoption of
python3 as the executable name, regardless of Py2 support.